Assignment_4 Part_4
Added a lot of things but forgot to do the commit as the code was not working properly. after that i added the class EvilCircle with extend shape. also added comments to every single line. also did even listener part again.

Changed the CheckBounds again with the requirements asked in the rubric.

used the collision detect again so that we can check the collision of the balls which goes through all the balls.

I did the collision the first one was for the color change and the second one was for the ball size.