Assignment_4 (Part_4)
After that, I created an array to hold balls. And for that, I have used the while loop.

then used the const random for the random size of the ball and that random is used for generating the random size of the ball using width and size.

after that, I used EvilCircle instance with the starting point 50.

used const and created the 'P' element to display the score. And then set initial text content to display the total number of balls. and then Append the score element to the document's body.

used function to run the animation loop. I have done the Draw, update, and check collision for each ball on the window.

EvilCircle used for draw, check bounds, and check collision.

then at last we have used repetition of the loop
with the variable name requestAnimationFrame(loop). Request the next frame to keep the animation loop running

and at last the loop begins while calling the loop function.