Assignment_4 (Part_2)
used the if loop for the toggling between the dark and light mode based on the respective class.

Also we have used an attribute so that if the button is in the dark mode it will switch to the light mode. Moreover, applied the dark overlay to the pictures. using the rgba.