Assignment_4 (Part_2)
the else block is used if the button is in light mode, it switch to the dark mode and similarly used the attribute with the parameters class and dark.

used the button content darken for the dark. Moreover, used the overlay style bg color rgba with value 0 to remove the overlay.